feat: add COM to deployment map filter
